Changes were made
There is now a try on the import...just seeing how that goes.

Also added a multilist box widget. It has test data in there at the moment. 

Added Labels to display for DICOMEcho depending on the result.

Also have checks for blank entry fields on the add device button.